+++ src/documentation/changelog.md
@@ -18,4 +18,5 @@ All notable changes to the Tarvo Lilja Portfolio Website will be documented in t
   - Tentatively finish Hero/Intro-section. Moving on to 'Education' next!
 
 - 27.11.2025
-  - Start work on Education-section
+  - Start work on Education-section.
+  - Finalize Education-section layout (Vertical top-to-bottom, alignment: Year ascending)